Add lint and husky, fix lint errors
