+++ src/App.css
@@ -1,5 +1,6 @@
 .App {
   text-align: center;
+  display: flex;
 }
 
 .App-logo {
@@ -24,8 +25,14 @@
   color: white;
 }
 
-.App-link {
+.link {
   color: #61dafb;
+  text-decoration: none;
+}
+
+.link:hover {
+  color: blue;
+  text-decoration: none;
 }
 
 @keyframes App-logo-spin {
